feat: PC 뷰에서도 try on 아이콘 추가 #81
